Fix #1: generic field ordering
With Generics, fields are arbitrarily nested with the binary operator
(:*:). Hence we cannot assume a left fold as we did until now to compute
field offsets. With this patch, we parameterize most operations of
GCStorable with an initial offset. The initial offset is only set to 0
for the whole data structure: i.e. in CStorable default methods.